requirements.txt

# reading environment vars:
python-dotenv

# web application:
Flask==3.0.2 # ==2.1.1
# Werkzeug==3.0.1

# auth (login with google via oauth):
Authlib==1.3.0 # ==1.0.0

# database (google sheets):
#google-auth==2.28.2 # ==2.16.2
#google-auth-oauthlib==1.2.0 #==1.0.0
gspread==6.0.2
gspread_models==1.0.2

google-api-python-client # ==2.142.0


# production web server:
gunicorn

# testing:
pytest
